changing step structure trying not to nest components so uses can change to right page step more smoothly nesting steps or step state is hindering user xp creating conditions that are not smooth all nested state to now be a step enabling easier integration with zustland
