+++ components/addMealStepper/stepFour/StepFour.tsx
@@ -1,0 +1,6 @@
+import React from "react";
+import styles from "../addMealStepper.module.scss";
+
+export default function StepFour() {
+  return <section className={styles.step__three}>StepThree</section>;
+}
